mmc: host: Migrate NEMU SD host controller
This commit adapts the NEMU SD host controller driver from
https://github.com/OpenXiangShan/NEMU/blob/master/resource/sdcard/nemu.c
to the current v6.10 kernel. The driver has been updated to align with
the latest MMC host controller framework and coding standards.

Note that this commit does not apply the interrupt handling workaround
mentioned in the NEMU SD driver documentation:
https://github.com/OpenXiangShan/NEMU/tree/master/resource/sdcard/README.md